requirements.txt
django==5.1.3
djangorestframework==3.15.2
moviepy==2.0.0
openai==1.55.3
django-cors-headers==4.3.1
librosa==0.10.1
opencv-python==4.9.0.80
mediapipe==0.10.9